Resolve Jetty dependencies
