+++ src/server-java/testapp-interactive/.classpath
@@ -7,7 +7,7 @@
 	<classpathentry combineaccessrules="false" exported="true" kind="src" path="/client"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/app"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/webcontainer"/>
-	<classpathentry kind="lib" path="jars/jetty-servlet.jar"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/jetty"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/servlet-api"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
