+++ src/server-java/testapp-interactive/eclipse/TestAppRun.launch
@@ -12,6 +12,7 @@
 <booleanAttribute key="run_jetty_run.ENABLE_SSL_ATTR" value="false"/>
 <booleanAttribute key="run_jetty_run.IGNORE_SCAN_CLASS_WHEN_DEBUG_MODE" value="true"/>
 <stringAttribute key="run_jetty_run.JETTY_XML_PATH" value=""/>
+<stringAttribute key="run_jetty_run.KEYSTORE_ATTR" value=""/>
 <stringAttribute key="run_jetty_run.KEY_PWD_ATTR" value="changeit"/>
 <stringAttribute key="run_jetty_run.PORT_ATTR" value="8080"/>
 <stringAttribute key="run_jetty_run.PWD_ATTR" value="changeit"/>
@@ -19,7 +20,7 @@
 <stringAttribute key="run_jetty_run.SELECTED_JETTY_VERSION_ATTR" value="Jetty 8.1.2.v20120308"/>
 <booleanAttribute key="run_jetty_run.SHOW_ADVANCE_ATTR" value="false"/>
 <stringAttribute key="run_jetty_run.SSL_PORT_ATTR" value="8443"/>
-<stringAttribute key="run_jetty_run.WEBAPPDIR_ATTR" value="webapp"/>
+<stringAttribute key="run_jetty_run.WEBAPPDIR_ATTR" value="eclipse/webapp"/>
 <listAttribute key="run_jetty_run.WEB_CONTEXT_CUSTOM_CLASSPATH">
 <listEntry value="&lt;?xml version=&quot;1.0&quot; encoding=&quot;UTF-8&quot; standalone=&quot;no&quot;?&gt;&#13;&#10;&lt;runtimeClasspathEntry internalArchive=&quot;/client/tmp/eclipse_build_js&quot; path=&quot;3&quot; type=&quot;2&quot;/&gt;&#13;&#10;"/>
 </listAttribute>
