+++ src/server-java/testapp-interactive/lib/nextapp/echo/testapp/interactive/InteractiveServlet.java
@@ -29,17 +29,16 @@
 
 package nextapp.echo.testapp.interactive;
 
-import jetty.JettyWebSocket;
 import nextapp.echo.app.ApplicationInstance;
-import nextapp.echo.webcontainer.ApplicationWebSocket;
 import nextapp.echo.webcontainer.Service;
 import nextapp.echo.webcontainer.WebContainerServlet;
-import nextapp.echo.webcontainer.WebSocketConnectionHandler;
 import nextapp.echo.webcontainer.service.JavaScriptService;
 import nextapp.echo.webcontainer.service.StaticTextService;
 
 /**
  * Interactive Test Application <code>WebContainerServlet</code> implementation.
+ * using polling as push mechanism
+ * (use the JettyInteractiveServlet if you want to try with WebSocket implementation) 
  */
 public class InteractiveServlet extends WebContainerServlet {
 
@@ -65,16 +64,8 @@ public class InteractiveServlet extends WebContainerServlet {
         System.setProperty("echo.allowiecompression", "true");
     }
 
-    public static final WebSocketConnectionHandler wsHandler = new WebSocketConnectionHandler() {
-        @Override
-        public ApplicationWebSocket newApplicationWebSocket(ApplicationInstance applicationInstance) {
-            return new JettyWebSocket();
-        }
-    };
-
     public InteractiveServlet() {
         super();
-        setWebSocketConnectionHandler(wsHandler);
         if (USE_CUSTOM_WAIT_INDICATOR) {
             addInitScript(CUSTOM_WAIT_INDICATOR);
         }
@@ -87,7 +78,7 @@ public class InteractiveServlet extends WebContainerServlet {
      * @see nextapp.echo.webcontainer.WebContainerServlet#getInstanceMode()
      */
     public int getInstanceMode() {
-        return INSTANCE_MODE_SINGLE;
+        return INSTANCE_MODE_WINDOW;
     }
 
     /**
